// template/react-starter-kit/src/.generated/api/model/updateTermRequestDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateTermRequestDto {
  /** @maxLength 50 */
  version?: string;
  content?: string;
  /** @nullable */
  publishedAt?: string | null;
}